+++ NEWS.md
@@ -1,4 +1,4 @@
-# simmer.plot 0.1.8.9000
+# simmer.plot 0.1.9
 
 ## Minor changes and fixes